Float32

+++ Covid19ModelVN/src/scripts/include/experiments.jl
@@ -20,7 +20,7 @@ function experiment_covid19_data(loc::AbstractString, train_range::Day, forecast
     df[2:end, :confirmed] .= diff(df[!, :confirmed_total])
     # ensure data type stability
     datacols = names(df, Not(:date))
-    df[!, datacols] .= Float64.(df[!, datacols])
+    df[!, datacols] .= Float32.(df[!, datacols])
 
     if loc == Covid19ModelVN.LOC_CODE_VIETNAM || loc ∈ keys(Covid19ModelVN.LOC_NAMES_VN)
         # we considered 27th April 2021 to be the start of the outbreak in Vietnam
@@ -49,7 +49,7 @@ end
 function experiment_movement_range(loc::AbstractString, first_date::Date, last_date::Date)
     df = get_prebuilt_movement_range(loc)
     cols = ["all_day_bing_tiles_visited_relative_change", "all_day_ratio_single_tile_users"]
-    df[!, cols] .= Float64.(df[!, cols])
+    df[!, cols] .= Float32.(df[!, cols])
     # smooth out weekly seasonality
     moving_average!(df, cols, 7)
     return load_timeseries(TimeseriesConfig(df, "ds", cols), first_date, last_date)
@@ -57,7 +57,7 @@ end
 
 function experiment_social_proximity(loc::AbstractString, first_date::Date, last_date::Date)
     df, col = get_prebuilt_social_proximity(loc)
-    df[!, col] .= Float64.(df[!, col])
+    df[!, col] .= Float32.(df[!, col])
     # smooth out weekly seasonality
     moving_average!(df, col, 7)
     return load_timeseries(TimeseriesConfig(df, "date", [col]), first_date, last_date)
@@ -77,7 +77,7 @@ function experiment_SEIRD_initial_states(
     E0 = I0 * 5 # exposed individuals
     S0 = population - E0 - df_first_date.confirmed_total[1] # susceptible individuals
     # initial state
-    u0 = Float64[S0, E0, I0, R0, D0, N0, C0, T0]
+    u0 = Float32[S0, E0, I0, R0, D0, N0, C0, T0]
     vars = [5, 7]
     labels = ["deaths", "new cases"]
     return u0, vars, labels
@@ -130,12 +130,12 @@ end
 
 function setup_baseline(
     loc::AbstractString;
-    γ0::Float64,
-    λ0::Float64,
-    α0::Float64,
-    γ_bounds::Tuple{Float64,Float64},
-    λ_bounds::Tuple{Float64,Float64},
-    α_bounds::Tuple{Float64,Float64},
+    γ0::Float32,
+    λ0::Float32,
+    α0::Float32,
+    γ_bounds::Tuple{Float32,Float32},
+    λ_bounds::Tuple{Float32,Float32},
+    α_bounds::Tuple{Float32,Float32},
     train_range::Day,
     forecast_range::Day,
     loss_type::Symbol,
@@ -173,12 +173,12 @@ end
 
 function setup_fbmobility1(
     loc::AbstractString;
-    γ0::Float64,
-    λ0::Float64,
-    α0::Float64,
-    γ_bounds::Tuple{Float64,Float64},
-    λ_bounds::Tuple{Float64,Float64},
-    α_bounds::Tuple{Float64,Float64},
+    γ0::Float32,
+    λ0::Float32,
+    α0::Float32,
+    γ_bounds::Tuple{Float32,Float32},
+    λ_bounds::Tuple{Float32,Float32},
+    α_bounds::Tuple{Float32,Float32},
     train_range::Day,
     forecast_range::Day,
     loss_type::Symbol,
@@ -220,12 +220,12 @@ end
 
 function setup_fbmobility2(
     loc::AbstractString;
-    γ0::Float64,
-    λ0::Float64,
-    α0::Float64,
-    γ_bounds::Tuple{Float64,Float64},
-    λ_bounds::Tuple{Float64,Float64},
-    α_bounds::Tuple{Float64,Float64},
+    γ0::Float32,
+    λ0::Float32,
+    α0::Float32,
+    γ_bounds::Tuple{Float32,Float32},
+    λ_bounds::Tuple{Float32,Float32},
+    α_bounds::Tuple{Float32,Float32},
     train_range::Day,
     forecast_range::Day,
     social_proximity_lag::Day,
@@ -282,13 +282,13 @@ end
 
 function setup_fbmobility3(
     loc::AbstractString;
-    γ0::Float64,
-    λ0::Float64,
-    α0::Float64,
-    β_bounds::Tuple{Float64,Float64},
-    γ_bounds::Tuple{Float64,Float64},
-    λ_bounds::Tuple{Float64,Float64},
-    α_bounds::Tuple{Float64,Float64},
+    γ0::Float32,
+    λ0::Float32,
+    α0::Float32,
+    β_bounds::Tuple{Float32,Float32},
+    γ_bounds::Tuple{Float32,Float32},
+    λ_bounds::Tuple{Float32,Float32},
+    α_bounds::Tuple{Float32,Float32},
     train_range::Day,
     forecast_range::Day,
     social_proximity_lag::Day,
@@ -346,12 +346,12 @@ end
 
 function setup_fbmobility4(
     loc::AbstractString;
-    γ0::Float64,
-    λ0::Float64,
-    β_bounds::Tuple{Float64,Float64},
-    γ_bounds::Tuple{Float64,Float64},
-    λ_bounds::Tuple{Float64,Float64},
-    α_bounds::Tuple{Float64,Float64},
+    γ0::Float32,
+    λ0::Float32,
+    β_bounds::Tuple{Float32,Float32},
+    γ_bounds::Tuple{Float32,Float32},
+    λ_bounds::Tuple{Float32,Float32},
+    α_bounds::Tuple{Float32,Float32},
     train_range::Day,
     forecast_range::Day,
     social_proximity_lag::Day,
@@ -522,8 +522,8 @@ function experiment_run(
     batch_timestamp = Dates.format(now(), "yyyymmddHHMMSS")
 
     lk_eval = ReentrantLock()
-    minimizers = Vector{Float64}[]
-    final_losses = Float64[]
+    minimizers = Vector{Float32}[]
+    final_losses = Float32[]
     queue_eval = Tuple{String,Function,Vector{Int},String}[]
 
     runexp = function (loc)
